Bills should store proportion as float.

+++ db/migrate/20110908180146_create_bills.rb
@@ -2,7 +2,7 @@ class CreateBills < ActiveRecord::Migration
   def change
     create_table :bills do |t|
       t.integer :basket_id
-      t.integer :proportion
+      t.float :proportion
 
       t.timestamps
     end
